app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
chromadb
requests
google-generativeai